• Refactored autoConnect() • Refactored PLPageViewController
