+++ PLENConnect/Scenography/Button.swift
@@ -13,11 +13,11 @@ class Button: UIButton {
     override func awakeFromNib() {
         super.awakeFromNib()
     
-        self.backgroundColor = Constants.Color.PlenGreenDark
-        self.layer.borderColor = UIColor.white.cgColor
-        self.layer.borderWidth = 1.0
-        self.layer.cornerRadius = 4.0
-        self.layer.masksToBounds = true
+        backgroundColor = Constants.Color.PlenGreenDark
+        layer.borderColor = UIColor.white.cgColor
+        layer.borderWidth = 1.0
+        layer.cornerRadius = 4.0
+        layer.masksToBounds = true
     }
     
 }
